misc: Reorganize Angular components
