+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { LoggedInGuard } from './logged-in.guard';
-import { NotLoggedInGuard } from './not-logged-in.guard';
+import { LoggedInGuard } from './guards/logged-in.guard';
+import { NotLoggedInGuard } from './guards/not-logged-in.guard';
 
-import { HomePageComponent } from './home-page/home-page.component';
-import { LoginPageComponent } from './login-page/login-page.component';
-import { TypeShowPageComponent } from './type-show-page/type-show-page.component';
+import { HomePageComponent } from './pages/home-page/home-page.component';
+import { LoginPageComponent } from './pages/login-page/login-page.component';
+import { TypeShowPageComponent } from './pages/type-show-page/type-show-page.component';
 
 const routes: Routes = [
   {
