+++ src/app/pages/home-page/home-page.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-home-page',
   templateUrl: './home-page.component.html',
-  styleUrls: ['./home-page.component.scss'],
+  styleUrls: [],
 })
 export class HomePageComponent implements OnInit {
   constructor () { }
