+++ src/app/pages/type-show-page/type-show-page.component.html
